requirements.txt
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
kafka-python==2.2.15
pydicom==3.0.1
pynetdicom==3.0.3
requests==2.32.4
urllib3==2.5.0
flask
python-dotenv 
